bglib: convert uncaught exceptions to events
This does not solve the underlying problem of an unhandled exception in
the connection handler thread, but means that the outer app layer
actually receives an event, giving them a chance of taking action.

We only have a single string field, but the traceback is still
important, so include it anyway, as a string, it's better than nothing.

Signed-off-by: Karl Palsson <karlp@etactica.com>